+++ src/config/index.js
@@ -7,7 +7,7 @@ if (emotesPath.startsWith('.')) {
 }
 
 module.exports = {
-  version: '1.1.0',
+  version: '1.2.0',
   apiVersion: 2,
   publicUrl: process.env.EMOTE_SERVER_PUBLIC_URL || 'http://localhost',
   port: process.env.EMOTE_SERVER_PORT || 8000,